% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh_api_client.R
\name{api_get_patients}
\alias{api_get_patients}
\title{api_get_patients}
\usage{
api_get_patients(base_url, token, num, num_type, only_num = FALSE,
  count = FALSE)
}
\value{
a num_temp pointing to the list of patients matchted
}
\description{
api_get_patients
}
