% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh_api_client.R
\name{api_get_token}
\alias{api_get_token}
\title{api_get_token}
\usage{
api_get_token(base_url, login, pwd)
}
\arguments{
\item{base_url:}{url for the api}
}
\value{
jwt token
}
\description{
api_get_token
}
