% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh_api_client.R
\name{api_match_patients}
\alias{api_match_patients}
\title{api_match_patients}
\usage{
api_match_patients(base_url, token, num, num_type, birth_range,
  concepts_range, n_match, match_save = FALSE, match_save_title = NULL)
}
\description{
api_match_patients
}
