% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh_api_client.R
\name{api_query}
\alias{api_query}
\title{api_query}
\usage{
api_query(base_url, endpoint, token, query_body = list(),
  METHOD = httr::POST)
}
\arguments{
\item{base_url:}{url for the api}

\item{token:}{jwt token (from get_token)}

\item{query_body:}{named list of parameters for the query}
}
\value{
result of the query (assumes a json format for the result)
}
\description{
api_query
}
