% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{create_cohorte}
\alias{create_cohorte}
\title{create_cohorte}
\usage{
create_cohorte(titre_cohorte, description_cohorte, username, config)
}
\arguments{
\item{titre_cohorte}{cohort title}

\item{description_cohorte}{description of the cohort}

\item{config}{a config environment created by the function getConfig.}
}
\value{
num_cohorte
}
\description{
For Dr.Warehouse
creates a cohort
}
