% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_num_temp_list}
\alias{get_num_temp_list}
\title{get_num_temp_list}
\usage{
get_num_temp_list(username = NULL, only_num = FALSE, config = NULL)
}
\arguments{
\item{username}{a username (string)}

\item{only_num}{set to `TRUE` if you want only a vector of num_temp numbers. Set to `FALSE`
if you also need the number of patients (default = `FALSE`)}

\item{config}{a config environment created by the function getConfig.}
}
\value{
a data.frame with the list of num_temps
}
\description{
For Dr.Warehouse
Get the list of recent queries results for a given username.
}
