% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_patients_counts_from_dwh}
\alias{get_patients_counts_from_dwh}
\title{get_patients_counts_from_dwh}
\usage{
get_patients_counts_from_dwh(config = config)
}
\arguments{
\item{config}{a config environment created by the function getConfig.}
}
\description{
For Dr. Warehouse
}
\details{
creates a table with the count of unique concepts per patients.
}
