% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{insert_patients_into_dwh_resultat}
\alias{insert_patients_into_dwh_resultat}
\title{insert_patients_into_dwh_resultat}
\usage{
insert_patients_into_dwh_resultat(patients, config)
}
\description{
insert_patients_into_dwh_resultat
}
