% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{insert_patients_into_query_result_instance}
\alias{insert_patients_into_query_result_instance}
\title{insert_patients_into_query_result_instance}
\usage{
insert_patients_into_query_result_instance(patients, config)
}
\arguments{
\item{patients}{a vector of patient_nums}

\item{config}{a config environment created by the function getConfig.}
}
\value{
result_id
}
\description{
For i2b2
inser patients in QUERY_RESULT_INSTANCE
}
