% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{match_patients_from_num}
\alias{match_patients_from_num}
\title{match_patients_from_num}
\usage{
match_patients_from_num(num = NULL, num_type = NULL, annee_range = NULL,
  count_range = NULL, n_match = NULL, match_save = FALSE,
  match_save_title = NULL, config = NULL)
}
\description{
match_patients_from_num
}
