% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{patients_subquery}
\alias{patients_subquery}
\title{patients_subquery}
\usage{
patients_subquery(num_type, backend)
}
\arguments{
\item{num_type}{one of c('num_temp', 'cohorte')}

\item{backend}{one of c('i2b2_oracle','i2b2_postgres', 'drwh_oracle' )}
}
\value{
SQL subquery
}
\description{
For i2b2 and Dr.Warehouse
Get the patients SQL subquery corresponding to the type cohort/num_temp and the backend Dr.Warehouse/i2b2
}
