% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{username_from_login}
\alias{username_from_login}
\title{username_from_login}
\usage{
username_from_login(login = NULL, config = NULL)
}
\arguments{
\item{login}{a string login}

\item{config}{a config environment created by the function getConfig.}
}
\value{
a character vector username.
}
\description{
For Dr.Warehouse
Get the internal username given a login
}
