% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh_api_client.R
\name{api_match_patients}
\alias{api_match_patients}
\title{api_match_patients}
\usage{
api_match_patients(base_url, token, num, num_type, annee_range, count_range,
  n_match)
}
\description{
api_match_patients
}
