% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfig.R
\name{getConfig}
\alias{getConfig}
\title{getConfig}
\usage{
getConfig(file_path)
}
\arguments{
\item{file_path}{Path to the config file}
}
\value{
config environment for DB connection
}
\description{
Load configuration parameters for DB connection
This file is a tab separated txt file of the form:
}
\details{
driverType     VALUE

driverPath     VALUE

connectPath VALUE

dbuser VALUE

dbpass VALUE
}
