% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{insert_patients_into_dwh_resultat}
\alias{insert_patients_into_dwh_resultat}
\title{insert_patients_into_dwh_resultat}
\usage{
insert_patients_into_dwh_resultat(patients, config)
}
\arguments{
\item{patients}{a vector of patient_nums}

\item{config}{a config environment created by the function getConfig.}
}
\value{
num_temp
}
\description{
For Dr.Warehouse
inser patients in DWH_RESULTAT
}
