% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{match_patient}
\alias{match_patient}
\title{match_patient}
\usage{
match_patient(num = NULL, num_type = NULL, sexe = NULL,
  annee_nais = NULL, annee_range = NULL, count_unique = NULL,
  count_range = NULL, n_match = NULL, config = NULL)
}
\description{
match_patient
}
