% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{match_patients_from_num}
\alias{match_patients_from_num}
\title{match_patients_from_num}
\usage{
match_patients_from_num(num = NULL, num_type = NULL, annee_range = NULL,
  count_range = NULL, n_match = NULL, match_save = FALSE,
  match_save_title = NULL, config = NULL)
}
\arguments{
\item{num}{the identifier of the cohort/num_temp}

\item{num_type}{type of num: one of c('num_temp', 'cohorte')}

\item{annee_range}{integer >= 0 to compute the range of authorized birth_years for matching.
the range will be: [annee_naiss - annee_range ; annee_nais + annee_range]}

\item{match_save}{For Dr.Warehouse (in i2b2 patient set is automatically saved). save the matched patients in a cohort (default = `FALSE`)}

\item{match_save_title}{title for the saved cohort.}

\item{config}{a config environment created by the function getConfig.}

\item{count_range.}{float > 0 to compute the range of authorized count_unique for matching.
the range will be: [count_unique - (count_unique * count_range) ; count_unique + (count_unique * count_range)]}

\item{n_match:}{number of patients to match}
}
\value{
the identifier of the matched cohort of patients
}
\description{
For Dr.Warehouse and i2b2
Match n (n_match) patients for each patient in a cohort/patient set.
}
