% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracleQuery.R
\name{oracleQuery}
\alias{oracleQuery}
\title{oracleQuery}
\usage{
oracleQuery(query, config = config, update = F, data = T)
}
\arguments{
\item{query}{SQL query in a character string}

\item{update}{set to `TRUE` if the query is an update. (default = `FALSE`)}

\item{data}{set to `TRUE` if the query returns data (default = `TRUE`)}

\item{configFile}{location of the config file (default = 'config')}
}
\value{
the result of the query (data.frame)
}
\description{
This function performs SQL queries to a database and returns the results.
}
